Autocomplete: restructure tree-sitter related modules
