+++ vscode/src/completions/tree-sitter/query-tests/test-data/blocks.ts
